+++ frontend/src/App.css
@@ -5,6 +5,7 @@
     --header-height: 70px;
     --default-color: #333333;
     --primary-color: #6200ea;
+    --error-color: red;
 }
 
 body {
