+++ frontend/src/components/CreateDataInvitation/CreateDataInvitation.styled.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const StyledParagraph = styled.p`
+    margin-bottom: 1rem;
+`;
