+++ frontend/src/pages/RestaurantsPage/RestaurantsPage.styled.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledErrorParagraph = styled.p`
+    font-size: .8rem;
+    color: var(--error-color);
+    margin-top: 1rem;
+`;
